Attempt to fix wasm-bindgen access error
