+++ src/lib.rs
@@ -106,11 +106,9 @@ mod optimizer;
 mod provider;
 #[cfg(feature = "ssr")]
 mod routes;
-mod picture;
 
 pub use image::*;
 pub use optimizer::{Filter, Resize, ResizeType};
-pub use picture::{Picture};
 #[cfg(feature = "ssr")]
 pub use optimizer::ImageOptimizer;
 pub use provider::*;
